+++ lizard_workspace/management/commands/sync_layers_fewsnorm.py
@@ -112,6 +112,7 @@ Example: bin/django sync_layers_fewsnorm --slug=<slug of existing Layer>
             new_layer.valid = True
             new_layer.is_local_server = layer.is_local_server
             new_layer.is_clickable = layer.is_local_server
+            new_layer.js_popup_class = layer.js_popup_class
             new_layer.save()
 
             new_layer.tags.add(tag)